refactor(types): consolidate shared table types and fix Funnel enum typo
- Move shared table types to @superset-ui/chart-controls:
  - BasicColorFormatterType
  - SortByItem
  - SearchOption
  - ServerPaginationData
  - TableColumnConfig
  - DataColumnMeta
  - CustomFormatter
- Update plugin-chart-table and plugin-chart-ag-grid-table to import
  from chart-controls and re-export for backwards compatibility

- Rename EchartsFunnelLabelTypeType to EchartsFunnelLabelType (fix typo)
- Update all references in types.ts, transformProps.ts, and controlPanel.tsx
